Response 3.1

+++ src/main/java/main/api/response/dto/PostDto.java
@@ -12,7 +12,7 @@ public class PostDto {
     private int id;
 
     @JsonProperty("timestamp")
-    private long timestamp;
+    private int timestamp;
 
     @JsonProperty("user")
     private PostUserDto user;
