+++ src/main/java/main/repositories/TagRepository.java
@@ -3,5 +3,5 @@ package main.repositories;
 import main.model.Tag;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TagRepository extends CrudRepository<Tag, Long> {
+public interface TagRepository extends CrudRepository<Tag, Integer> {
 }
